Function list improvements
Former-commit-id: b468cafe1cf8f766440f2ed0f3418d055c5f7f5c

+++ Source/simba.bitmap_misc.pas
@@ -114,7 +114,7 @@ begin
       Bitmap.SetSize(Area.Width, Area.Height);
 
       PixelOffset := FileHeader.bfOffset;
-      ScanLineSize := ((Header.Width * Header.BitCount) + 31) div 32 * 4;
+      ScanLineSize := Int64((Header.Width * Header.BitCount) + 31) div 32 * 4;
       Buffer := GetMem(ScanLineSize);
       Index := 0;
